requirements.txt
Flask==2.3.3
requests==2.31.0
#pandas==2.0.3
pandas==1.5.3
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
tenacity==8.2.3
numpy==1.23.5
